# remove unused items in `import {..}` only when required (for better LSP experience)