--- a/build/package.template.yaml
+++ b/build/package.template.yaml
@@ -14,6 +14,7 @@ default-extensions:
   - StrictData
 
 dependencies:
+  - aeson
   - ansi-terminal
   - async
   - base