## before
  - `neut format-source` tracks all the imported files recursively and remove unused items in `import {..}` automatically
  - `textDocument/formatting` == (format + remove used items in `import {..}`)

## after
  - `neut format-source` simply formats given files
  - `neut format-source --minimize-imports` formats given files and removes unused items in `import {..}`
  - `textDocument/formatting` == (format)
  - Added a new code action for (format + remove used items in `import {..}`)

## benchmark

```sh
# format ← (used by textDocument/formatting)
❯ hyperfine "neut format-source source/neural.nt"
Benchmark 1: neut format-source source/neural.nt
  Time (mean ± σ):      33.6 ms ±   1.3 ms    [User: 14.9 ms, System: 15.8 ms]
  Range (min … max):    30.4 ms …  36.1 ms    82 runs

# format + remove used items in `import {..}` ← (used by textDocument/codeAction)
❯ hyperfine "neut format-source --minimize-imports source/neural.nt"
Benchmark 1: neut format-source --minimize-imports source/neural.nt
  Time (mean ± σ):     135.7 ms ±   7.4 ms    [User: 245.9 ms, System: 74.6 ms]
  Range (min … max):   124.3 ms … 148.8 ms    19 runs
```